splitting the kifi pane out into its own script
