+++ packrat/scripts/keeper_scout.js
@@ -6,8 +6,7 @@ function logEvent() {  // parameters defined in main.js
   api.port.emit("log_event", Array.prototype.slice.call(arguments));
 }
 
-var session,
-  tags = [];
+var session, tags = [];
 var tile = tile || function() {  // idempotent for Chrome
   'use strict';
   log("[keeper_scout]", location.hostname)();
@@ -23,10 +22,10 @@ var tile = tile || function() {  // idempotent for Chrome
   api.port.on({
     session_change: onSessionChange,
     open_to: function(o) {
-      keeper("showPane", o.trigger, o.locator, o.redirected);
+      loadAndDo('pane', 'show', o.trigger, o.locator, o.redirected);
     },
-    button_click: keeper.bind(null, "togglePane", "button"),
-    auto_show: keeper.bind(null, "show", "auto"),
+    button_click: loadAndDo.bind(null, 'pane', 'toggle', 'button'),
+    auto_show: loadAndDo.bind(null, 'slider2', 'show', 'auto'),
     init: function(o) {
       var pos = o.position;
       if (pos) {
@@ -59,7 +58,7 @@ var tile = tile || function() {  // idempotent for Chrome
       }
     },
     keepers: function(o) {
-      setTimeout(keeper.bind(null, "showKeepers", o.keepers, o.otherKeeps), 3000);
+      setTimeout(loadAndDo.bind(null, 'slider2', 'showKeepers', o.keepers, o.otherKeeps), 3000);
     },
     counts: function(counts) {
       tile && updateCounts(counts);
@@ -77,7 +76,7 @@ var tile = tile || function() {  // idempotent for Chrome
             log("[onScroll]", Math.round(hSeen / hPage * 10000) / 100, ">", r[1], "% and", hPage, ">", r[0] * hViewport, "?")();
             if (hPage > r[0] * hViewport && hSeen > (r[1] / 100) * hPage) {
               log("[onScroll] showing")();
-              keeper("show", "scroll");
+              loadAndDo('slider2', 'show', 'scroll');
             }
           }
         });
@@ -108,11 +107,11 @@ var tile = tile || function() {  // idempotent for Chrome
         e.preventDefault();
         break;
       case 77: // m
-        keeper("togglePane", "key", "/messages");
+        loadAndDo('pane', 'toggle', 'key', '/messages');
         e.preventDefault();
         break;
       case 79: // o
-        keeper("togglePane", "key", "/notices");
+        loadAndDo('pane', 'toggle', 'key', '/notices');
         e.preventDefault();
         break;
       }
@@ -136,20 +135,22 @@ var tile = tile || function() {  // idempotent for Chrome
     });
   }
 
-  function keeper() {  // gateway to slider2.js
-    var args = Array.prototype.slice.apply(arguments);
+  function loadAndDo(name, methodName) {  // gateway to slider2.js
     if (session === undefined) {  // not yet initialized
+      var args = Array.prototype.slice.call(arguments);
       args.unshift(null);
-      whenSessionKnown.push(Function.bind.apply(keeper, args));
+      whenSessionKnown.push(Function.bind.apply(loadAndDo, args));
     } else if (!session) {
       toggleLoginDialog();
     } else {
-      api.require("scripts/slider2.js", function() {
-        if (onScroll && name != "showKeepers") {
-          document.removeEventListener("scroll", onScroll);
+      var args = Array.prototype.slice.call(arguments, 2);
+      api.require('scripts/' + name + '.js', function() {
+        if (onScroll && name !== 'showKeepers') {
+          document.removeEventListener('scroll', onScroll);
           onScroll = null;
         }
-        slider2[args.shift()].apply(slider2, args);
+        var o = window[name];
+        o[methodName].apply(o, args);
       });
     }
   }
@@ -184,7 +185,7 @@ var tile = tile || function() {  // idempotent for Chrome
   tile["kifi:position"] = positionTile;
   tile.addEventListener("mouseover", function(e) {
     if (e.target === tileCount || tileCard.contains(e.target)) {
-      keeper("show", "tile");
+      loadAndDo('slider2', 'show', 'tile');
     }
   });
 
@@ -194,7 +195,7 @@ var tile = tile || function() {  // idempotent for Chrome
   document.addEventListener("keydown", onKeyDown, true);
 
   function onResize() {
-    if (document.documentElement.classList.contains("kifi-with-pane")) return;
+    if (paneCall('showing')) return;
     clearTimeout(onResize.t);  // throttling tile repositioning
     onResize.t = setTimeout(positionTile, 50);
   }
@@ -228,8 +229,13 @@ var tile = tile || function() {  // idempotent for Chrome
         tile.remove();
       }
     }
-    if (window.slider2) {
-      slider2.hidePane();
+    paneCall('hide');
+  }
+
+  function paneCall(methodName) {
+    var pane = window.pane;
+    if (pane) {
+      return pane[methodName]();
     }
   }
